have the register_maps library of LUTs compiled separately (#233)
* have the register_maps library of LUTs compiled separately

this makes it so we only have to wait for that long compilation when one
of the YAML files changes which /hopefully/ doesn't happen too
frequently

* test: build econ maps in static get rather than initializer_list

* also have roc maps avoid initializer lists

* move ugly register_maps CMake code into subdirectory CMakeLists

mainly trying to keep the ugliness out of view in the central CMakeLists. The central CMakeLists can then just use the 'register_maps' target like any other target.

* remove any reference to register_maps in public includes

this makes it installable

* Apply clang-format --style=Google

* remove commented out code

* remove unnecessary target

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>