dont pass std::shared_ptr by reference (#289)
This defeats the purpose of a std::shared_ptr which only knows to
increment its reference count when it gets copied. When passing by
reference, the reference counter does not get incremented, the original
copies of std::shared_ptr are deleted and the object is destructed
leading to a pointer that references invalid memory.